Fix opening/closing for awning
