+++ homeassistant/components/overkiz/cover_entities/awning.py
@@ -11,7 +11,12 @@ from homeassistant.components.cover import (
     CoverEntityFeature,
 )
 
-from .generic_cover import COMMANDS_STOP, OverkizGenericCover
+from .generic_cover import (
+    COMMANDS_CLOSE,
+    COMMANDS_OPEN,
+    COMMANDS_STOP,
+    OverkizGenericCover,
+)
 
 
 class Awning(OverkizGenericCover):
@@ -64,3 +69,43 @@ class Awning(OverkizGenericCover):
     async def async_close_cover(self, **kwargs: Any) -> None:
         """Close the cover."""
         await self.executor.async_execute_command(OverkizCommand.UNDEPLOY)
+
+    @property
+    def is_opening(self) -> bool | None:
+        """Return if the cover is opening or not."""
+
+        if self.assumed_state:
+            return None
+
+        if self.is_running(COMMANDS_OPEN):
+            return True
+
+        # Check if cover is moving based on current state
+        is_moving = self.device.states.get(OverkizState.CORE_MOVING)
+        current_closure = self.device.states.get(OverkizState.CORE_DEPLOYMENT)
+        target_closure = self.device.states.get(OverkizState.CORE_TARGET_CLOSURE)
+
+        if not is_moving or not current_closure or not target_closure:
+            return None
+
+        return cast(int, current_closure.value) < cast(int, target_closure.value)
+
+    @property
+    def is_closing(self) -> bool | None:
+        """Return if the cover is closing or not."""
+
+        if self.assumed_state:
+            return None
+
+        if self.is_running(COMMANDS_CLOSE):
+            return True
+
+        # Check if cover is moving based on current state
+        is_moving = self.device.states.get(OverkizState.CORE_MOVING)
+        current_closure = self.device.states.get(OverkizState.CORE_DEPLOYMENT)
+        target_closure = self.device.states.get(OverkizState.CORE_TARGET_CLOSURE)
+
+        if not is_moving or not current_closure or not target_closure:
+            return None
+
+        return cast(int, current_closure.value) > cast(int, target_closure.value)
